(chores) camel-jms: consolidate duplicate test code
