+++ components/camel-jms/src/test/java/org/apache/camel/component/jms/AbstractPersistentJMSTest.java
@@ -17,6 +17,10 @@
 
 package org.apache.camel.component.jms;
 
+import javax.jms.ConnectionFactory;
+
+import org.apache.camel.CamelContext;
+import org.apache.camel.test.infra.activemq.common.ConnectionFactoryHelper;
 import org.apache.camel.test.infra.activemq.services.ActiveMQService;
 import org.apache.camel.test.infra.activemq.services.ActiveMQServiceFactory;
 import org.apache.camel.test.junit5.CamelTestSupport;
@@ -25,9 +29,25 @@ import org.junit.jupiter.api.Tags;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
+import static org.apache.camel.component.jms.JmsComponent.jmsComponentAutoAcknowledge;
+
 @Tags({ @Tag("jms") })
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public abstract class AbstractPersistentJMSTest extends CamelTestSupport {
     @RegisterExtension
     public static ActiveMQService service = ActiveMQServiceFactory.createPersistentVMService();
+
+    @Override
+    protected CamelContext createCamelContext() throws Exception {
+        CamelContext camelContext = super.createCamelContext();
+
+        createConnectionFactory(camelContext);
+
+        return camelContext;
+    }
+
+    protected void createConnectionFactory(CamelContext camelContext) {
+        ConnectionFactory connectionFactory = ConnectionFactoryHelper.createConnectionFactory(service);
+        camelContext.addComponent("activemq", jmsComponentAutoAcknowledge(connectionFactory));
+    }
 }
